File the feeds are updated the menu will show 'Loading...' and an indicator message is shown when the loading is started.
All feed information is stored in a sqlite3 database. Therefore the config and the cache file are no longer necessary. All posts of a feed are stored forever, currently there is no option for deleting old posts from the database. Feed titles can be changed in the config window.
The database will be created if the doesn't exist including the necessary tables. The database is stored in the app config dir.